[runtime] Move interpreter code from shrimp_vm
- Code is moved to runtime/interpreter
- Adjust includes and cmake
