+++ runtime/CMakeLists.txt
@@ -4,14 +4,16 @@ add_library(runtime STATIC)
 add_library(shrimp::runtime ALIAS runtime)
 
 add_subdirectory(common)
-
-# Add runtime_runtime include and src
-add_subdirectory(runtime)
-# Add runtime_shrimp_vm include and src
 add_subdirectory(shrimp_vm)
+
+# Add runtime::runtime include and src
+add_subdirectory(runtime)
+# Add runtime::interpreter include and src
+add_subdirectory(interpreter)
 
 target_link_libraries(runtime
 PUBLIC
     shrimp::common
-    shrimp::runtime_common
+    shrimp::runtime::common
+    shrimp::runtime::shrimp_vm
 )
